Permission of the file changed.
In my system the permission given to the file was only of read and write beacuse
I had created this file using the tocuh command. So, in order to execute this file
and the result in shell I gave the execute permission to this file. The command
to give that permission is "chmod "+x" video2.sh".