Video 7 file created.
In this code file, I had written a program that will write some data at the end
if the file. First it will check that whether the file exist or not, if yes than
it will check whether that file has write permission or not, if both if are
true then through command line we can add text to that file which will appear at the
end the file.